[chore] #118 plan 서버통신 파일 수정
