--- app/src/main/java/org/sopt/pingle/data/datasource/remote/PlanMeetingRemoteDataSource.kt
@@ -1,8 +1,0 @@
-package org.sopt.pingle.data.datasource.remote
-
-import org.sopt.pingle.data.model.remote.request.RequestPlanMeetingDto
-import org.sopt.pingle.util.base.NullableBaseResponse
-
-interface PlanMeetingRemoteDataSource {
-    suspend fun postPlanMeeting(teamId: Int, requestPlanMeetingDto: RequestPlanMeetingDto): NullableBaseResponse<Unit?>
-}
