+++ app/src/main/java/org/sopt/pingle/data/repository/PlanRepositoryImpl.kt
@@ -4,7 +4,9 @@ import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import org.sopt.pingle.data.datasource.remote.PlanRemoteDataSource
+import org.sopt.pingle.data.mapper.toRequestPlanMeetingDto
 import org.sopt.pingle.domain.model.PlanLocationEntity
+import org.sopt.pingle.domain.model.PlanMeetingEntity
 import org.sopt.pingle.domain.repository.PlanRepository
 
 class PlanRepositoryImpl @Inject constructor(
@@ -19,4 +21,17 @@ class PlanRepositoryImpl @Inject constructor(
             }
             emit(result.getOrThrow())
         }
+
+    override suspend fun postPlanMeeting(
+        teamId: Int,
+        planMeetingEntity: PlanMeetingEntity
+    ): Flow<Unit?> = flow {
+        val result = runCatching {
+            planRemoteDataSource.postPlanMeeting(
+                teamId = teamId,
+                requestPlanMeetingDto = planMeetingEntity.toRequestPlanMeetingDto()
+            ).data
+        }
+        emit(result.getOrThrow())
+    }
 }
