+++ app/src/main/java/org/sopt/pingle/di/RepositoryModule.kt
@@ -4,19 +4,17 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
 import org.sopt.pingle.data.repository.AuthRepositoryImpl
 import org.sopt.pingle.data.repository.JoinGroupRepositoryImpl
 import org.sopt.pingle.data.repository.MapRepositoryImpl
 import org.sopt.pingle.data.repository.PingleRepositoryImpl
-import org.sopt.pingle.data.repository.PlanMeetingRepositoryImpl
 import org.sopt.pingle.data.repository.PlanRepositoryImpl
 import org.sopt.pingle.domain.repository.AuthRepository
 import org.sopt.pingle.domain.repository.JoinGroupRepository
 import org.sopt.pingle.domain.repository.MapRepository
 import org.sopt.pingle.domain.repository.PingleRepository
-import org.sopt.pingle.domain.repository.PlanMeetingRepository
 import org.sopt.pingle.domain.repository.PlanRepository
+import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -36,10 +34,6 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindsPingleRepository(pingleRepositoryImpl: PingleRepositoryImpl): PingleRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindsPlanMeetingRepository(planMeetingRepositoryImpl: PlanMeetingRepositoryImpl): PlanMeetingRepository
 
     @Binds
     @Singleton
