--- app/src/main/java/org/sopt/pingle/domain/repository/PlanMeetingRepository.kt
@@ -1,11 +1,0 @@
-package org.sopt.pingle.domain.repository
-
-import kotlinx.coroutines.flow.Flow
-import org.sopt.pingle.domain.model.PlanMeetingEntity
-
-interface PlanMeetingRepository {
-    suspend fun postPlanMeeting(
-        teamId: Int,
-        planMeetingEntity: PlanMeetingEntity
-    ): Flow<Unit?>
-}
