+++ app/src/main/java/org/sopt/pingle/domain/usecase/PostPlanMeetingUseCase.kt
@@ -2,13 +2,13 @@ package org.sopt.pingle.domain.usecase
 
 import kotlinx.coroutines.flow.Flow
 import org.sopt.pingle.domain.model.PlanMeetingEntity
-import org.sopt.pingle.domain.repository.PlanMeetingRepository
+import org.sopt.pingle.domain.repository.PlanRepository
 
 class PostPlanMeetingUseCase(
-    private val planMeetingRepository: PlanMeetingRepository
+    private val planRepository: PlanRepository
 ) {
     suspend operator fun invoke(teamId: Int, planMeetingEntity: PlanMeetingEntity): Flow<Unit?> =
-        planMeetingRepository.postPlanMeeting(
+        planRepository.postPlanMeeting(
             teamId = teamId,
             planMeetingEntity = planMeetingEntity
         )
